cleanup 32 bit API

+++ src/libponyc/ast/stringtab.c
@@ -23,7 +23,7 @@ typedef struct stringtab_entry_t
 
 static size_t stringtab_hash(stringtab_entry_t* a)
 {
-  return hash_block(a->str, a->len);
+  return (size_t)hash_block(a->str, a->len);
 }
 
 static bool stringtab_cmp(stringtab_entry_t* a, stringtab_entry_t* b)
